pyproject.toml
[tool.poetry]
name = "uber_popug_task_tracker"
version = "0.1.0"
description = ""
authors = ["Nadezhda Lavrenteva <lavrhope@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
pre-commit = "^3.3.3"
ruff = "^0.0.280"
black = "^23.7.0"
isort = "^5.12.0"
uvicorn = "^0.23.2"
pydantic = {extras = ["dotenv"], version = "^2.1.1"}
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
aiopg = {version = "^1.4.0", extras = ["sa"]}
ipdb = "^0.13.13"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
bcrypt = "4.0.1"  # не обновлять!
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
